get-unique-ids-new -> get-unique-ids
