+++ editor/src/core/model/get-unique-ids.benchmark.ts
@@ -3,42 +3,10 @@ import type { ParseSuccess } from 'utopia-shared/src/types'
 import type { ProjectContentTreeRoot } from '../../components/assets'
 import { modifyParseSuccessAtPath } from '../../components/editor/store/editor-state'
 import { parseProjectContents } from '../../sample-projects/sample-project-utils.test-utils'
-import { getAllUniqueUids, getAllUniqueUidsInnerOld, getFilePathForUid } from './get-unique-ids'
+import { getAllUniqueUids, getFilePathForUid } from './get-unique-ids'
 import { LargeHydrogenProject } from './test-large-persistent-model.test-utils'
-import {
-  clearCachedUidsPerFileForTests,
-  getUniqeUidsMapping,
-  getUniqueUidsMappingInner,
-  lookupFilePathForUid,
-} from './get-unique-ids-new'
 
 export async function benchmarkGetUniqueUids(): Promise<void> {
-  await Benny.suite(
-    'COLD RUN collect a mapping of uid -> filename',
-    Benny.add('OLD getUniqueUids', async function setup() {
-      const parsedProjectContents = parseProjectContents(LargeHydrogenProject.projectContents)
-
-      return function benchmark() {
-        const mapping = getAllUniqueUidsInnerOld(parsedProjectContents).filePathToUids
-        const result = getFilePathForUid(mapping, '8d4c5e9960f004aebf00a97dabe7cd80')
-      }
-    }),
-    Benny.add('getUniqueUids FIFTH', async function setup() {
-      const parsedProjectContents = parseProjectContents(LargeHydrogenProject.projectContents)
-
-      return function benchmark() {
-        clearCachedUidsPerFileForTests()
-        const mapping = getUniqueUidsMappingInner(parsedProjectContents)
-        const result = lookupFilePathForUid(
-          mapping.filePathToUids,
-          '8d4c5e9960f004aebf00a97dabe7cd80',
-        )
-      }
-    }),
-    Benny.cycle(),
-    Benny.complete(),
-  )
-
   await Benny.suite(
     'ONE FILE CHANGED, memoized collection of uid -> filename',
     Benny.add('OLD getUniqueUids', async function setup() {
@@ -48,18 +16,6 @@ export async function benchmarkGetUniqueUids(): Promise<void> {
         parsedProjectContents = changeOneFileInProjectContents(parsedProjectContents)
         const mapping = getAllUniqueUids(parsedProjectContents).filePathToUids
         const result = getFilePathForUid(mapping, '8d4c5e9960f004aebf00a97dabe7cd80')
-      }
-    }),
-    Benny.add('getUniqueUids FIFTH', async function setup() {
-      let parsedProjectContents = parseProjectContents(LargeHydrogenProject.projectContents)
-
-      return function benchmark() {
-        parsedProjectContents = changeOneFileInProjectContents(parsedProjectContents)
-        const mapping = getUniqeUidsMapping(parsedProjectContents)
-        const result = lookupFilePathForUid(
-          mapping.filePathToUids,
-          '8d4c5e9960f004aebf00a97dabe7cd80',
-        )
       }
     }),
     Benny.cycle(),
@@ -80,24 +36,6 @@ export async function benchmarkGetUniqueUids(): Promise<void> {
         const oldResult = getFilePathForUid(oldMapping, '8d4c5e9960f004aebf00a97dabe7cd80')
       }
     }),
-    Benny.add('getUniqueUids FIFTH', async function setup() {
-      const parsedProjectContents = parseProjectContents(LargeHydrogenProject.projectContents)
-
-      return function benchmark() {
-        const changedProjectContents = changeOneFileInProjectContents(parsedProjectContents)
-        const newMapping = getUniqeUidsMapping(parsedProjectContents)
-        const newResult = lookupFilePathForUid(
-          newMapping.filePathToUids,
-          '8d4c5e9960f004aebf00a97dabe7cd80',
-        )
-
-        const oldMapping = getUniqeUidsMapping(parsedProjectContents)
-        const oldResult = lookupFilePathForUid(
-          oldMapping.filePathToUids,
-          '8d4c5e9960f004aebf00a97dabe7cd80',
-        )
-      }
-    }),
     Benny.cycle(),
     Benny.complete(),
   )
@@ -110,17 +48,6 @@ export async function benchmarkGetUniqueUids(): Promise<void> {
 
       return function benchmark() {
         const result = getFilePathForUid(mapping, '8d4c5e9960f004aebf00a97dabe7cd80')
-      }
-    }),
-    Benny.add('getUniqueUids FIFTH', async function setup() {
-      let parsedProjectContents = parseProjectContents(LargeHydrogenProject.projectContents)
-      const mapping = getUniqeUidsMapping(parsedProjectContents)
-
-      return function benchmark() {
-        const result = lookupFilePathForUid(
-          mapping.filePathToUids,
-          '8d4c5e9960f004aebf00a97dabe7cd80',
-        )
       }
     }),
     Benny.cycle(),
