+++ editor/src/core/workers/parser-printer/uid-fix.spec.ts
@@ -668,10 +668,7 @@ function checkUIDValues([first, second]: [JSXElementChild, JSXElementChild]): bo
   return (
     isEmptyObject(firstUIDsResult.duplicateIDs) &&
     isEmptyObject(secondUIDsResult.duplicateIDs) &&
-    intersection([
-      new Set(getAllUniqueUidsFromLookup(firstUIDsResult.filePathToUids)),
-      new Set(getAllUniqueUidsFromLookup(secondUIDsResult.filePathToUids)),
-    ]).size === 0
+    intersection([firstUIDsResult.allUids, secondUIDsResult.allUids]).size === 0
   )
 }
